+++ src/Parameter.h
@@ -41,7 +41,7 @@ class Parameter: public Variable {
                 unsigned int mode,
                 enum si_datatype_t dtype,
                 unsigned int ndims = 1,
-                const size_t *dim = 0);
+                const unsigned int *dim = 0);
 
         virtual ~Parameter();
