+++ src/Signal.h
@@ -40,7 +40,7 @@ class Signal: public Variable {
                 const char *path,
                 enum si_datatype_t dtype,
                 unsigned int ndims = 1,
-                const size_t *dim = 0);
+                const unsigned int *dim = 0);
 
         virtual ~Signal();
