+++ src/lib/Parameter.h
@@ -40,7 +40,7 @@ class Parameter: public HRTLab::Parameter {
                 enum si_datatype_t dtype,
                 void *addr,
                 unsigned int ndims = 1,
-                const size_t *dim = 0);
+                const unsigned int *dim = 0);
 
         ~Parameter();
